:recycle: moved processes into server initialization. 	  moved spool folder creation into server initialization. 	  created a Queue file for the queue struct.
